requirements.txt
Django==2.2.1
djangorestframework==3.9.4
Markdown==3.1.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2019.1
sqlparse==0.3.0
